# apps/onboarding-ms/src/main/docs/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Onboarding API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ],
...
  "paths" : {
...
    "/v1/onboarding" : {
      "get" : {
        "tags" : [ "Onboarding Controller" ],
        "summary" : "Retrieve paged onboardings with optional filters and sorting.",
        "description" : "The API retrieves paged onboarding using optional filter, order by descending creation date",
        "operationId" : "getOnboardingWithFilter",
        "parameters" : [ {
          "name" : "from",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "institutionId",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "onboardingId",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "default" : "0",
            "type" : "integer"
          }
        }, {
          "name" : "productId",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "schema" : {
            "format" : "int32",
            "default" : "20",
            "type" : "integer"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "subunitCode",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "taxCode",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OnboardingGetResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "Onboarding Controller" ],
        "summary" : "Default onboarding for GSP/SA/AS institutions, saves user data, and triggers async onboarding activities.",
        "description" : "Perform default onboarding request, it is used for GSP/SA/AS institution type.Users data will be saved on personal data vault if it doesn't already exist.At the end, function triggers async activities related to onboarding based on institution type.",
        "operationId" : "onboarding",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OnboardingDefaultRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OnboardingResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
...
    "/v1/onboarding/completion" : {
      "post" : {
        "tags" : [ "Onboarding Controller", "internal-v1" ],
        "summary" : "Complete onboarding request and set status to COMPLETED.",
        "description" : "Perform onboarding as /onboarding but completing the onboarding request to COMPLETED phase.",
        "operationId" : "onboardingCompletion",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OnboardingDefaultRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OnboardingResponseV1"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/v1/onboarding/institutionOnboardings" : {
      "get" : {
        "tags" : [ "support", "internal-v1", "Onboarding" ],
        "summary" : "Get onboardings by institution taxCode, subunitCode, origin, or originId.",
        "description" : "Returns onboardings record by institution taxCode/subunitCode/origin/originId",
        "operationId" : "onboardingInstitutionUsingGET",
        "parameters" : [ {
          "name" : "origin",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "originId",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/OnboardingStatus"
          }
        }, {
          "name" : "subunitCode",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "taxCode",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/OnboardingResponseV1"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ],
        "x-legacy-api" : "true"
      }
    },
...
    "/v1/onboarding/pa/completion" : {
      "post" : {
        "tags" : [ "Onboarding Controller", "internal-v1" ],
        "summary" : "Complete PA onboarding request and set status to COMPLETED.",
        "description" : "Perform onboarding as /onboarding/pa but completing the onboarding request to COMPLETED phase.",
        "operationId" : "onboardingPaCompletion",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OnboardingPaRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OnboardingResponseV1"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
...
    "/v1/onboarding/pg/completion" : {
      "post" : {
        "tags" : [ "Onboarding Controller", "internal-pnpg" ],
        "summary" : "Complete PG onboarding request on PNPG domain and set status to COMPLETED.",
        "description" : "Perform onboarding as /onboarding/psp but completing the onboarding request to COMPLETED phase.",
        "operationId" : "onboardingPgCompletion",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OnboardingPgRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OnboardingResponseV1"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
...
    "/v1/onboarding/psp/completion" : {
      "post" : {
        "tags" : [ "Onboarding Controller", "internal-v1" ],
        "summary" : "Complete PSP onboarding request and set status to COMPLETED.",
        "description" : "Perform onboarding as /onboarding/psp but completing the onboarding request to COMPLETED phase.",
        "operationId" : "onboardingPspCompletion",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OnboardingPspRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OnboardingResponseV1"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
...
    "/v1/onboarding/{onboardingId}/pending" : {
      "get" : {
        "tags" : [ "Onboarding Controller" ],
        "summary" : "Get pending onboarding by ID.",
        "description" : "Returns an onboarding record by its ID only if its status is PENDING. This feature is crucial for ensuring that the onboarding process can be completed only when the onboarding status is appropriately set to PENDING.",
        "operationId" : "getOnboardingPending",
        "parameters" : [ {
          "name" : "onboardingId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OnboardingGet"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
...
    "/v2/onboarding/completion" : {
      "post" : {
        "tags" : [ "Onboarding Controller", "internal-v1" ],
        "summary" : "Complete onboarding request and set status to COMPLETED.",
        "description" : "Perform onboarding as /onboarding but completing the onboarding request to COMPLETED phase.",
        "operationId" : "onboardingCompletion",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OnboardingDefaultRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OnboardingResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/v2/onboarding/institutionOnboardings" : {
      "get" : {
        "tags" : [ "support", "internal-v1", "Onboarding" ],
        "summary" : "Get onboardings by institution taxCode, subunitCode, origin, or originId.",
        "description" : "Returns onboardings record by institution taxCode/subunitCode/origin/originId",
        "operationId" : "onboardingInstitutionUsingGET",
        "parameters" : [ {
          "name" : "origin",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "originId",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/OnboardingStatus"
          }
        }, {
          "name" : "subunitCode",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "taxCode",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/OnboardingResponse"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ],
        "x-legacy-api" : "true"
      }
    },
    "/v2/onboarding/pa/completion" : {
      "post" : {
        "tags" : [ "Onboarding Controller", "internal-v1" ],
        "summary" : "Complete PA onboarding request and set status to COMPLETED.",
        "description" : "Perform onboarding as /onboarding/pa but completing the onboarding request to COMPLETED phase.",
        "operationId" : "onboardingPaCompletion",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OnboardingPaRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OnboardingResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/v2/onboarding/pg/completion" : {
      "post" : {
        "tags" : [ "Onboarding Controller", "internal-pnpg" ],
        "summary" : "Complete PG onboarding request on PNPG domain and set status to COMPLETED.",
        "description" : "Perform onboarding as /onboarding/psp but completing the onboarding request to COMPLETED phase.",
        "operationId" : "onboardingPgCompletion",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OnboardingPgRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OnboardingResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/v2/onboarding/psp/completion" : {
      "post" : {
        "tags" : [ "Onboarding Controller", "internal-v1" ],
        "summary" : "Complete PSP onboarding request and set status to COMPLETED.",
        "description" : "Perform onboarding as /onboarding/psp but completing the onboarding request to COMPLETED phase.",
        "operationId" : "onboardingPspCompletion",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OnboardingPspRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OnboardingResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "BillingRequest" : {
        "type" : "object",
        "properties" : {
          "vatNumber" : {
            "type" : "string"
          },
          "recipientCode" : {
            "type" : "string"
          },
          "publicServices" : {
            "type" : "boolean"
          }
        }
      },
...
      "InstitutionType" : {
        "enum" : [ "PA", "PG", "GSP", "SA", "PT", "SCP", "PSP", "AS", "REC", "CON", "PRV", "GPU" ],
        "type" : "string"
      },
      "LocalDateTime" : {
        "format" : "date-time",
        "type" : "string",
        "example" : "2022-03-10T12:15:50"
      },
...
      "OnboardingDefaultRequest" : {
        "required" : [ "productId", "users", "institution" ],
        "type" : "object",
        "properties" : {
          "productId" : {
            "minLength" : 1,
            "type" : "string"
          },
          "users" : {
            "minItems" : 1,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserRequest"
            }
          },
          "aggregates" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AggregateInstitutionRequest"
            }
          },
          "isAggregator" : {
            "type" : "boolean"
          },
          "pricingPlan" : {
            "type" : "string"
          },
          "signContract" : {
            "type" : "boolean"
          },
          "institution" : {
            "$ref" : "#/components/schemas/InstitutionBaseRequest"
          },
          "billing" : {
            "$ref" : "#/components/schemas/BillingRequest"
          },
          "additionalInformations" : {
            "$ref" : "#/components/schemas/AdditionalInformationsDto"
          },
          "gpuData" : {
            "$ref" : "#/components/schemas/GPUData"
          }
        }
      },
      "OnboardingGet" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "productId" : {
            "type" : "string"
          },
          "workflowType" : {
            "type" : "string"
          },
          "institution" : {
            "$ref" : "#/components/schemas/InstitutionResponse"
          },
          "users" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserResponse"
            }
          },
          "pricingPlan" : {
            "type" : "string"
          },
          "billing" : {
            "$ref" : "#/components/schemas/BillingResponse"
          },
          "signContract" : {
            "type" : "boolean"
          },
          "additionalInformations" : {
            "$ref" : "#/components/schemas/AdditionalInformationsDto"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/LocalDateTime"
          },
          "updatedAt" : {
            "$ref" : "#/components/schemas/LocalDateTime"
          },
          "expiringDate" : {
            "$ref" : "#/components/schemas/LocalDateTime"
          },
          "status" : {
            "type" : "string"
          },
          "userRequestUid" : {
            "type" : "string"
          },
          "reasonForReject" : {
            "type" : "string"
          },
          "attachments" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "OnboardingGetResponse" : {
        "type" : "object",
        "properties" : {
          "count" : {
            "format" : "int64",
            "type" : "integer"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/OnboardingGet"
            }
          }
        }
      },
...
      "OnboardingPaRequest" : {
        "required" : [ "productId", "users", "institution" ],
        "type" : "object",
        "properties" : {
          "productId" : {
            "minLength" : 1,
            "type" : "string"
          },
          "users" : {
            "minItems" : 1,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserRequest"
            }
          },
          "aggregates" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AggregateInstitutionRequest"
            }
          },
          "isAggregator" : {
            "type" : "boolean"
          },
          "pricingPlan" : {
            "type" : "string"
          },
          "signContract" : {
            "type" : "boolean"
          },
          "institution" : {
            "$ref" : "#/components/schemas/InstitutionBaseRequest"
          },
          "billing" : {
            "$ref" : "#/components/schemas/BillingPaRequest"
          },
          "additionalInformations" : {
            "$ref" : "#/components/schemas/AdditionalInformationsDto"
          }
        }
      },
      "OnboardingPgRequest" : {
        "required" : [ "productId", "users", "taxCode", "origin", "digitalAddress" ],
        "type" : "object",
        "properties" : {
          "productId" : {
            "minLength" : 1,
            "type" : "string"
          },
          "institutionType" : {
            "$ref" : "#/components/schemas/InstitutionType"
          },
          "users" : {
            "minItems" : 1,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserRequest"
            }
          },
          "taxCode" : {
            "type" : "string"
          },
          "businessName" : {
            "type" : "string"
          },
          "origin" : {
            "$ref" : "#/components/schemas/Origin"
          },
          "digitalAddress" : {
            "type" : "string"
          }
        }
      },
      "OnboardingPspRequest" : {
        "required" : [ "productId", "users", "institution" ],
        "type" : "object",
        "properties" : {
          "productId" : {
            "minLength" : 1,
            "type" : "string"
          },
          "users" : {
            "minItems" : 1,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserRequest"
            }
          },
          "aggregates" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AggregateInstitutionRequest"
            }
          },
          "isAggregator" : {
            "type" : "boolean"
          },
          "pricingPlan" : {
            "type" : "string"
          },
          "signContract" : {
            "type" : "boolean"
          },
          "institution" : {
            "$ref" : "#/components/schemas/InstitutionPspRequest"
          },
          "billing" : {
            "$ref" : "#/components/schemas/BillingRequest"
          }
        }
      },
      "OnboardingResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "productId" : {
            "type" : "string"
          },
          "workflowType" : {
            "type" : "string"
          },
          "institution" : {
            "$ref" : "#/components/schemas/InstitutionResponse"
          },
          "pricingPlan" : {
            "type" : "string"
          },
          "users" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserOnboardingResponse"
            }
          },
          "billing" : {
            "$ref" : "#/components/schemas/BillingResponse"
          },
          "status" : {
            "type" : "string"
          },
          "additionalInformations" : {
            "$ref" : "#/components/schemas/AdditionalInformationsDto"
          },
          "userRequestUid" : {
            "type" : "string"
          },
          "isAggregator" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/LocalDateTime"
          },
          "updatedAt" : {
            "$ref" : "#/components/schemas/LocalDateTime"
          }
        }
      },
      "OnboardingResponseV1" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "productId" : {
            "type" : "string"
          },
          "workflowType" : {
            "type" : "string"
          },
          "institution" : {
            "$ref" : "#/components/schemas/InstitutionResponse"
          },
          "pricingPlan" : {
            "type" : "string"
          },
          "users" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserOnboardingResponseV1"
            }
          },
          "billing" : {
            "$ref" : "#/components/schemas/BillingResponse"
          },
          "status" : {
            "type" : "string"
          },
          "additionalInformations" : {
            "$ref" : "#/components/schemas/AdditionalInformationsDto"
          },
          "userRequestUid" : {
            "type" : "string"
          },
          "isAggregator" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/LocalDateTime"
          },
          "updatedAt" : {
            "$ref" : "#/components/schemas/LocalDateTime"
          }
        }
      },
      "OnboardingStatus" : {
        "enum" : [ "REQUEST", "TOBEVALIDATED", "PENDING", "COMPLETED", "FAILED", "REJECTED", "DELETED" ],
        "type" : "string"
      },
...
      "UserOnboardingResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "role" : {
            "$ref" : "#/components/schemas/PartyRole"
          },
          "productRole" : {
            "type" : "string"
          },
          "userMailUuid" : {
            "type" : "string"
          }
        }
      },
...
      "UserRequest" : {
        "type" : "object",
        "properties" : {
          "taxCode" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "surname" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "role" : {
            "$ref" : "#/components/schemas/PartyRole"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "http",
        "description" : "Authentication",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}